API/Section9/1_recap_of_code/start/requirements.txt
Flask-JWT
-Extended
Flask-RESTful

Flask-SQLAlchemy